Added Rules to compendium
